Implement boosting algorithm using corrective terms
This commit adds an additional version of the refinement algorithm that uses the information from the refinement metric to adjust specific edges using the corrective terms instead of cutting them. Just like the previous version, the refinement will first perform a reference run, which should be configured to not use any domain knowledge. Then another run (called the original run) will take place which should use domain knowledge. The model trained in this run will be refined over the specified number of refinement runs. In order for this method to work, the ModelConfig needs to have use_corrective_terms set to True. Other parameters behave the same as in the original version of the refinement algorithm. The additional new parameter corrective_factor in RefinementConfig controls the decay of the attention value in terms of the refinement metric, where a value > 1 means a stronger correction and a value < 1 means a weaker correction.