Allow ids of existing runs as originals during refinement
This commit allows specifying an existing run to use as the original run during refinement instead of requiring a new run to be performed. The trained weights of the original run are restored before the refinement runs take place and therefore the model architecture needs to be weight-compatible with the given existing run.