Allow specific model layers to be frozen during refinement
This commit adds parameters to control if and when the RNN, embedding and activation layers of the model are frozen (i.e. made untrainable) during the refinement. It is possible to specify a list of boolean values for the freeze_rnn_sequence, freeze_embeddings_sequence and freeze_activiation_sequence parameters in the RefinementConfig to indicate if a specific layer should be frozen in the corresponding iteration. If no value is specified for an iteration, the last specified value is used.