Compare against last non-noisy run during generation
This commit changes the high-level generation to compare against the last run without noise instead of comparing against a reference (without domain knowledge). It is no longer possible to add noise to the original run, but it is now possible to specify an id to use. There is no longer a reference run and any parameters for it will be ignored.